Almost working. Home position added
