+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ArmShoulderPositions.java
@@ -2,22 +2,22 @@ package org.firstinspires.ftc.teamcode;
 
 public  enum ArmShoulderPositions {
 
-    INTAKE(0,0.2,0,1.0),
-    LEVEL1(1100,0.6,0.5,0.5),
-    LEVEL2(860,0.6,1,0),
-    LEVEL3(600,0.6,1,0),
-    TSE_DROP(635,0.6,1,0),
-    GROUND_PICKUP(1000,0.6,0.7,0.3);
+    INTAKE(50,0.1,0.23,0.77),
+    LEVEL1(1100,0.4,0.2,0.8),
+    LEVEL2(860,0.4,0.4,0.6),
+    LEVEL3(600,0.4,0.75,0.25),
+    TSE_DROP(600,0.4,0.54,0.46),
+    GROUND_PICKUP(1000,0.4,0.5,0.5),
+    HOME(100,0.4,0.23,0.77);
 
-
-    private final int armTarget;
+    private  final int armTarget;
     private final double maxPower;
     private final double leftWristServoPosition;
     private final double rightWristServoPosition;
 
 
-    private ArmShoulderPositions(int motorPosition, double maxPower, double leftWristServoPosition, double rightWristServoPosition) {
-        this.armTarget = motorPosition;
+    private ArmShoulderPositions(int armTarget, double maxPower, double leftWristServoPosition, double rightWristServoPosition) {
+        this.armTarget = armTarget;
         this.maxPower = maxPower;
         this.leftWristServoPosition = leftWristServoPosition;
         this.rightWristServoPosition = rightWristServoPosition;
